// src/app/modules/Address/address.interface.ts
import { Document, Types } from 'mongoose';

export interface IAddress extends Document {
  user: Types.ObjectId;
  label?: string;
  line1: string;
  line2?: string;
  city: string;
  state?: string;
  postalCode?: string;
  country?: string;
  location?: {
    type: 'Point';
    coordinates: [number, number];
  };
  isDefault: boolean;
  createdAt: Date;
  updatedAt: Date;
}